requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
backoff==1.10.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
fastapi==0.78.0
ffmpy==0.3.0
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
gdown==4.5.1
gradio==3.0.22
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
kiwisolver==1.4.3
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mdit-py-plugins==0.3.0
mdurl==0.1.1
monotonic==1.6
multidict==6.0.2
numpy==1.23.0
opencv-contrib-python==4.1.2.30
orjson==3.7.6
packaging==21.3
pandas==1.4.3
paramiko==2.11.0
Pillow==9.2.0
pip==21.2.4
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
scipy==1.8.1
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
torch==1.8.2+cpu
torchvision==0.9.2+cpu
tqdm==4.64.0
typing_extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.9
uvicorn==0.18.2
wheel==0.37.1
yarl==1.7.2
